image-processor/requirements.txt
boto3==1.34.69
Pillow==10.2.0
